% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoconnexr-package.R
\docType{package}
\name{geoconnexr-package}
\alias{geoconnexr}
\alias{geoconnexr-package}
\title{geoconnexr: Geoconnex discovery and watershed data snapshots}
\description{
\code{geoconnexr} provides typed, versioned foundations for discovering and
packaging water data described through the Geoconnex ecosystem. The
package is currently experimental and its network-facing APIs are being
developed behind fixture-backed contracts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ksonda.github.io/geoconnexr/}
  \item \url{https://github.com/ksonda/geoconnexr}
  \item Report bugs at \url{https://github.com/ksonda/geoconnexr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Onda \email{konda@lincolninst.edu} [copyright holder]

}
\keyword{internal}
