% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{gx_validate_comid}
\alias{gx_validate_comid}
\title{Validate NHDPlus COMIDs}
\usage{
gx_validate_comid(x)
}
\arguments{
\item{x}{A character vector containing positive integer identifiers.}
}
\value{
\code{x}, invisibly, after validation.
}
\description{
Validate NHDPlus COMIDs
}
